+++ runtime/doc/various.txt
@@ -456,7 +456,8 @@ N  *+textprop*		|text-properties|
 N  *+timers*		the |timer_start()| function
 N  *+title*		Setting the window 'title' and 'icon'
 N  *+toolbar*		|gui-toolbar|
-N  *+user_commands*	User-defined commands. |user-commands|
+T  *+user_commands*	User-defined commands. |user-commands|
+			Always enabled since 8.1.1210.
 B  *+vartabs*		Variable-width tabstops. |'vartabstop'|
 N  *+viminfo*		|'viminfo'|
    *+vertsplit*		Vertically split windows |:vsplit|; Always enabled
