+++ src/Make_dice.mak
@@ -83,6 +83,7 @@ SRC = \
 	term.c \
 	ui.c \
 	undo.c \
+	usercmd.c \
 	userfunc.c \
 	window.c \
 	version.c
@@ -144,6 +145,7 @@ OBJ =	o/arabic.o \
 	o/term.o \
 	o/ui.o \
 	o/undo.o \
+	o/usercmd.o \
 	o/userfunc.o \
 	o/window.o \
 	$(TERMLIB)
@@ -287,6 +289,8 @@ o/termlib.o:	termlib.c $(SYMS)
 o/ui.o: 	ui.c	$(SYMS)
 
 o/undo.o: 	undo.c	$(SYMS)
+
+o/usercmd.o: 	usercmd.c  $(SYMS)
 
 o/userfunc.o: 	userfunc.c  $(SYMS)
